# [Enhance] change max_transmit_batched_bytes from 64KB to 256KB